requirements.txt
Flask
Flask-Login
Flask-WTF
Flask-SQLAlchemy
Flask-Babel
Flask-Migrate
Werkzeug
gunicorn
requests
psycopg2-binary
bootstrap-flask
python-dotenv
python-dateutil
